requirements.txt
# Core dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
transformers>=4.37.0
torch>=2.1.0
pillow>=10.0.0
numpy>=1.26.0
scipy>=1.12.0
opencv-python-headless>=4.9.0.80
decord>=0.6.0
av>=11.0.0
python-multipart>=0.0.6
aiosqlite>=0.19.0
pydantic-settings>=2.1.0
pyrogram>=2.0.106,<2.1
tgcrypto>=1.2.5
ruamel.yaml>=0.18.10

# Frozen CLIP training
deepspeed>=0.14.0
